Moved selection box rendering to the entity move tool
